Added Uma2 Configuration
