+++ src/Services/KeycloakService.php
@@ -3,18 +3,17 @@
 namespace Vizir\KeycloakWebGuard\Services;
 
 use Exception;
-use GuzzleHttp\ClientInterface;
-use GuzzleHttp\Exception\GuzzleException;
 use Illuminate\Support\Arr;
+use GuzzleHttp\ClientInterface;
 use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Config;
-use Illuminate\Support\Facades\Route;
+use GuzzleHttp\Exception\GuzzleException;
 use Vizir\KeycloakWebGuard\Auth\KeycloakAccessToken;
-use Vizir\KeycloakWebGuard\Auth\Guard\KeycloakWebGuard;
 
 class KeycloakService
 {
+    use OpenIdConfiguration, Uma2Configuration;
+
     /**
      * The Session key for token
      */
@@ -52,20 +51,6 @@ class KeycloakService
      * @var string
      */
     protected $clientSecret;
-
-    /**
-     * Keycloak OpenId Configuration
-     *
-     * @var array
-     */
-    protected $openid;
-
-    /**
-     * Keycloak OpenId Cache Configuration
-     *
-     * @var array
-     */
-    protected $cacheOpenid;
 
     /**
      * CallbackUrl
@@ -123,6 +108,10 @@ class KeycloakService
 
         if (is_null($this->cacheOpenid)) {
             $this->cacheOpenid = Config::get('keycloak-web.cache_openid', false);
+        }
+
+        if (is_null($this->cacheUma2)) {
+            $this->cacheOpenid = Config::get('keycloak-web.cache_uma2', false);
         }
 
         if (is_null($this->callbackUrl)) {
@@ -476,66 +465,6 @@ class KeycloakService
     protected function getState()
     {
         return $this->state;
-    }
-
-    /**
-     * Return a value from the Open ID Configuration
-     *
-     * @param  string $key
-     * @return string
-     */
-    protected function getOpenIdValue($key)
-    {
-        if (! $this->openid) {
-            $this->openid = $this->getOpenIdConfiguration();
-        }
-
-        return Arr::get($this->openid, $key);
-    }
-
-    /**
-     * Retrieve OpenId Endpoints
-     *
-     * @return array
-     */
-    protected function getOpenIdConfiguration()
-    {
-        $cacheKey = 'keycloak_web_guard_openid-' . $this->realm . '-' . md5($this->baseUrl);
-
-        // From cache?
-        if ($this->cacheOpenid) {
-            $configuration = Cache::get($cacheKey, []);
-
-            if (! empty($configuration)) {
-                return $configuration;
-            }
-        }
-
-        // Request if cache empty or not using
-        $url = $this->baseUrl . '/realms/' . $this->realm;
-        $url = $url . '/.well-known/openid-configuration';
-
-        $configuration = [];
-
-        try {
-            $response = $this->httpClient->request('GET', $url);
-
-            if ($response->getStatusCode() === 200) {
-                $configuration = $response->getBody()->getContents();
-                $configuration = json_decode($configuration, true);
-            }
-        } catch (GuzzleException $e) {
-            $this->logException($e);
-
-            throw new Exception('[Keycloak Error] It was not possible to load OpenId configuration: ' . $e->getMessage());
-        }
-
-        // Save cache
-        if ($this->cacheOpenid) {
-            Cache::put($cacheKey, $configuration);
-        }
-
-        return $configuration;
     }
 
     /**
